feat: interface ParseTransition
The interface StateTransition for the parse operation result
allowing future conversion to an object.

The SimpleStateTrnasition does not care indexes.

The IterableStateTransition does require new state
to be iterable parse result with index. The interface
also assumes that the default result type is an array
of elements, and require the result is iterable of
elements.

The IterableParseState no longer require result type to be
Iterable<ELEMENT>, but defaults to Iterable<ELEMENT>.

modified:   src/stateparser.mjs
- faet: interface ParseTransition
- feat return type of StateParser and IterableStateParser
changed to the StateTransition.
- feat: interface SimpleStateTransition
- feat: interface IterableStateTransition